requirements.txt
Django==2.2
django-extensions==3.1.0
djangorestframework==3.9.2
gunicorn==20.0.4
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
